Replace imagine::open() method with passthru alt
Instead of using imagine::open() in end process (when display the image to
user) we're not using more robus approach (simply outputing what whe have
to the end user).
This was done because we're already using cached path, thus we suppose to
have correct format in place.
The only case is when given file format is not supported and we need to
output file as is.