Make use of not_found images parameter
Make use of not_found images parameter in Twig extension and templating
helper.
Now we'll try to resolve NULL uri/path to a valid image.
This also was done to prevent filtering not_found images set in
configuration (now we need to set an absolute path/uri to image that
should be used if an empty value is provided by argument.